Continued modifications based on IDL/python comparisons.
