+++ src/pyspextool/combine/__init__.py
@@ -1,2 +1,4 @@
-from .combine_spectra import combine_spectra
+from .combine import combine
 from .load_spectra import load_spectra
+from .scale_spectra import scale_spectra
+from .combine_spectra import combine_spectra
