+++ src/pyspextool/combine/config.py
@@ -1,27 +1,1 @@
-state = {'headers':1,
-         'final_napertures':1,
-         'module':1,
-         'napertures':1,
-         'nfiles':1,
-         'norders':1,
-         'npixels':1,
-         'nspectra':1,
-         'orders':1,
-         'wavelengths':1,
-         'input_files':1,
-         'intensity':1,         
-         'intensities':1,
-         'uncertainty':1,         
-         'uncertainties':1,
-         'bitmask':1,
-         'bitmasks':1,
-         'scales':1,
-         'xlabel':'',
-         'xunits':''}         
-
-load = {'files':1, 'output_name':1, 'input_path':1, 'output_path':1,
-        'scale_spectra':1, 'scale_order':1, 'scale_range':1,
-        'scale_range_fraction':1, 'correct_spectral_shape':1, 'statistic':1,
-        'robust_sigma':1, 'qafile':True, 'qaplot':True, 'qaplotsize':(6,6),
-        'line_width':0.5, 'verbose':True}
-
+state = {}
