+++ src/pyspextool/extract/background_subtraction.py
@@ -1,9 +1,15 @@
 import numpy as np
+import numpy.typing as npt
 
 from pyspextool.io.check import check_parameter
 
 
-def median_1dxd(image, edgecoeffs, xranges, var=None, ybuffer=1):
+def median_1dxd(image:npt.ArrayLike,
+                edgecoeffs:npt.ArrayLike,
+                xranges:npt.ArrayLike,
+                var:npt.ArrayLike=None,
+                ybuffer:int=1):
+    
     """
     To background subtraction a 1DXD image.
 
@@ -15,7 +21,7 @@ def median_1dxd(image, edgecoeffs, xranges, var=None, ybuffer=1):
     image : ndarray
         (nrows, ncols) image.
 
-    edgecoeffs : array_like of float
+    edgecoeffs : ndarray
         (norders,2,ncoeffs) array giving the polynomial 
         coefficients delineating the top and bottom of each order.  
         edgecoeffs[0,0,:] gives the coefficients for the bottom of 
@@ -23,16 +29,16 @@ def median_1dxd(image, edgecoeffs, xranges, var=None, ybuffer=1):
         edgecoeffs[0,1,:] gives the coefficients for the top of said 
         order.  
 
-    xranges : array_like of int, [norders,2] 
+    xranges : ndarray of int
         An (norders,2) array giving the column numbers over which to 
         search.  sranges[0,0] gives the starting column number for 
         the first order and sranges[0,1] gives the end column number 
         for the first order.
 
-    var : ndarray, optional
+    var : ndarray, default None
         (nrows, ncols) variance image.    
 
-    ybuffer : int, default=1
+    ybuffer : int, default 1
         The number of native pixels from the top and bottom of the slit to
         avoid during the operation.  Useful to account for the fact that
         the drop-off in intensity at the edge of the slit is not a
@@ -40,8 +46,14 @@ def median_1dxd(image, edgecoeffs, xranges, var=None, ybuffer=1):
 
     Returns
     -------
-    Later
+    ndarray, optional ndarray
+    The background subtracted image where the median intensity is subtracted
+    from each column in each row.
 
+    If var is not None, then (1.4826 * MAD) ** 2, where MAD is the median
+    absolute deviation given by median(abs(med-data)), is added to each
+    column.  
+    
     """
 
     #
@@ -101,8 +113,8 @@ def median_1dxd(image, edgecoeffs, xranges, var=None, ybuffer=1):
             # Do the error propagation if need be
 
             if var is not None:
-                mad = np.median(np.abs(image[bot_idx:top_idx, j] - med))
-                var[bot_idx:top_idx, j] += (1.4826 * mad) ** 2
+                mad = np.median(np.abs(image[bot_idx:top_idx+1, j] - med))
+                var[bot_idx:top_idx+1, j] += (1.4826 * mad) ** 2
 
     # Return the results
