+++ src/pyspextool/instrument_data/uspex_dir/uspex.dat
@@ -81,7 +81,6 @@ COMBINE_KEYWORD=SLTW_ARC
 COMBINE_KEYWORD=SLTH_PIX
 COMBINE_KEYWORD=SLTW_PIX
 COMBINE_KEYWORD=RP
-COMBINE_KEYWORD=APOSO*
 COMBINE_KEYWORD=BDPXFIX 
 COMBINE_KEYWORD=THRESH  
 COMBINE_KEYWORD=OPTEXT  
@@ -94,6 +93,7 @@ COMBINE_KEYWORD=LXUNITS
 COMBINE_KEYWORD=LYUNITS 
 COMBINE_KEYWORD=LXLABEL 
 COMBINE_KEYWORD=LYLABEL
+COMBINE_KEYWORD=LULABEL
 COMBINE_KEYWORD=HISTORY
 
 #
